refactor: extract shared session cookie verification helper
